uint16_t -> uint32_t vanwege baudrate 115200

+++ Inc/UART.h
@@ -9,7 +9,7 @@
 
 #include "stm32f0xx_hal.h"
 
-int initUART(uint16_t _baudRate);
+int initUART(uint32_t _baudRate);
 int transmit(uint8_t *data, int size);
 
 #endif /* UART_H_ */
